% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{BBC News Summaries}
\format{
A data frame. 100 observations of 3 variables:
\describe{
  \item{article}{the written article}
  \item{summary}{the human written summary}
  \item{textrank}{the textrank generated summary}
}
}
\source{
\url{https://www.kaggle.com/pariza/bbc-news-summary}
}
\usage{
data
}
\description{
A subset of a dataset containing 100 BBC News articles, alongside a human written summary and one generated using the Textrank algorithm
}
\keyword{datasets}
