% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_measure.R
\name{f_measure}
\alias{f_measure}
\title{F-Measure}
\usage{
f_measure(recall, precision)
}
\arguments{
\item{recall}{a single recall value}

\item{precision}{a single precision value}
}
\value{
The F1-score
}
\description{
For given recall and precision values, computes the F1-score
}
