% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_rouge_n.R
\name{multi_rouge_n}
\alias{multi_rouge_n}
\title{Rouge N with multiple reference summaries}
\usage{
multi_rouge_n(candidate, references, n)
}
\arguments{
\item{candidate}{the candidate (usually machine generated) summary}

\item{references}{a character vector of reference (usually human written) summaries}

\item{n}{the number of grams to be considered.  eg. n = 2 considers bigrams, generating a ROUGE-2 score.}
}
\value{
The ROUGE-N score
}
\description{
Takes a candidate summary, more than one reference summary, and a single n value.  Returns a ROUGE score for the given N.
}
