% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rouge_n.R
\name{rouge_n}
\alias{rouge_n}
\title{Rouge-N: N-gram Co-occurrence Statistics}
\usage{
rouge_n(candidate, reference, n = 1)
}
\arguments{
\item{candidate}{the candidate (usually machine generated) summary}

\item{reference}{the reference (usually human written) summary}

\item{n}{the number of grams to be considered, default 1.}
}
\value{
a dataframe, consisting of a single observation and three variables: recall, precision, and f_measure.
}
\description{
Takes a candidate summary, reference summary, and n value, returning the associated ROUGE-N recall, precision, and F values.
}
