% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rouge_s.R
\name{rouge_s}
\alias{rouge_s}
\title{Rouge-S: Skip-Bigram Co-occurence Statistics}
\usage{
rouge_s(candidate, reference, k = 1, beta = 1)
}
\arguments{
\item{candidate}{the candidate (usually machine generated) summary}

\item{reference}{the reference (usually human written) summary}

\item{k}{the maximum skip distance, default 1}

\item{beta}{a measure of the importance of precision in relation to recall in the F-measure, default 1—implying equal importance.}
}
\value{
a dataframe, consisting of a single observation and three variables: recall, precision, and f_measure.
}
\description{
Takes a candidate summary and reference summary, returning the associated ROUGE-S recall, precision, and F values.
}
