% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rouge_n.R
\name{rouge_n}
\alias{rouge_n}
\title{Rouge N}
\usage{
rouge_n(candidate, reference, n)
}
\arguments{
\item{candidate}{the candidate (usually machine generated) summary}

\item{reference}{the reference (usually human written) summary}

\item{n}{the number of grams to be considered.  eg. n = 2 considers bigrams, generating a ROUGE-2 score.}
}
\value{
The ROUGE-N score
}
\description{
Takes a candidate summary, reference summary, and n value.  Returns a ROUGE score for the given N.
}
